requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.0.1
httpx==0.28.1
google-genai==1.51.0
pydantic==2.12.4
python-multipart==0.0.17
reflex==0.6.7
pypdf==5.1.0
cryptography==43.0.3


